Template TreeConstruction::calculateBoundingBoxOfTheScene() on the type of the primitives
Move the definition to the declaration file and inline the member function.
